Transfered class Task defenition and declaration from threadpool  files to task files
