+++ CMakeLists.txt
@@ -9,10 +9,12 @@ project(${PROJECT_NAME} VERSION ${PROJECT_VESRION})
 
 set(SOURCES src/jsonParser.cpp src/config.cpp src/threadpool.cpp src/manager.cpp
             src/queue.cpp src/builder.cpp src/recorder.cpp
-        src/commonStructures.cpp)
+        src/commonStructures.cpp
+        src/task.cpp)
 set(HEADERS include/config.hpp include/commonStructures.hpp
         include/jsonParser.hpp include/recorder.hpp include/threadpool.hpp
-        include/manager.hpp include/interfaces.hpp include/queue.hpp include/builder.hpp)
+        include/manager.hpp include/interfaces.hpp include/queue.hpp include/builder.hpp
+        include/task.hpp)
 
 add_executable(${PROJECT_NAME} src/main.cpp ${SOURCES} ${HEADERS})
 target_include_directories(${PROJECT_NAME} PUBLIC include)
